refactor: 메인 페이지 feature 폴더 구조 변경
