+++ features/main/components/calendar/index.ts
@@ -1,3 +1,3 @@
+export * from '../../types/calendar';
 export * from './molecules';
 export * from './organisms';
-export * from './types';
