+++ features/main/components/calendar/molecules/calendar.tsx
@@ -4,8 +4,8 @@ import { useSetAtom } from 'jotai';
 import { useEffect } from 'react';
 
 import { LoadingArea } from '@/components/atoms';
+import { useCalendarRendaringData } from '@/features/main/hooks';
 import { useCalendarData } from '@/hooks';
-import { useCalendarRendaringData } from '@/hooks/use-calendar-rendering-data';
 import { calendarSwimCountAtom } from '@/store';
 import { css } from '@/styled-system/css';
 import { flex } from '@/styled-system/patterns';
