+++ features/main/components/main-tab/index.ts
@@ -1,0 +1,1 @@
+export * from './main-tab';
