+++ features/main/components/time-line/atoms/swim-record-chart.tsx
@@ -1,9 +1,8 @@
+import { StrokeInfo } from '@/features/main/types';
 import { StrokeName } from '@/features/record-detail';
 import { css, cva } from '@/styled-system/css';
 import { flex, grid } from '@/styled-system/patterns';
 import { getSwimColor } from '@/utils';
-
-import { StrokeInfo } from '../types';
 
 interface SwimRecordChartProps {
   width: number;
