+++ features/main/components/time-line/index.ts
@@ -1,3 +1,2 @@
 export * from './molecules';
 export * from './organisms';
-export * from './types';
