+++ features/main/components/time-line/organisms/time-line.tsx
@@ -2,13 +2,13 @@
 import { css, cx } from '@/styled-system/css';
 import { flex } from '@/styled-system/patterns';
 import { TimeLineCard } from '../molecules';
-import { TimeLineContent } from '../types';
 import { formatDateToKorean } from '@/utils';
 import { Fragment } from 'react';
 import { CardWrapper } from '../atoms';
 import { InfiniteScroller } from '@/components/molecules';
 import { SwimmerIcon } from '@/components/atoms';
 import { useTimeLineData } from '@/hooks';
+import { TimeLineContent } from '@/features/main/types';
 
 interface GroupedTimelineContents {
   date: string;
