--- features/main/main-tab/index.ts
@@ -1,1 +1,0 @@
-export { MainTab } from './main-tab';
